Build gflags and glog through CMake if not found in the system
This adds functionality to fetch gflags and glog from GitHub and build
them during the Caffe build. This happens only if a system-wide
installed version is not found (i.e., when find_package() fails).

If built as part of the Caffe build, gflags and glog are compiled as
position-independent static libraries. This avoids doing a system-wide
install of these libraries during the Caffe install target.